Module espresso_client::assets

The assets module.

This module defines Asset, AssetEditor, and Assets, which provide CURD (create, read, update, and delete) operations, with the use of KeyValueStore to control the assets resource.

Structs

Asset

An asset with its code as the primary key.

AssetEditor

An editor to create or update the asset or assets store.

Assets

Assets stored in an assets store.

Icon

A small icon to display with an asset in a GUI interface.

MintInfo

Information required to mint an asset.

VerifiedAssetLibrary

A library of assets, signed by a trusted party.

Type Definitions

AssetsStore