Module espresso_client::cli

The generic CAP Keystore frontend

This module “frontend” provides a framework for implementing command line interfaces for ledger-specific instantiations of the Keystore type. Similar to the Keystore framework itself, there is are traits which must be implemented to adapt this framework to a particular ledger type, after which the implementor gains access to the full Seahorse CLI implementation.

The CLI trait must be implemented for a particular ledger type and KeystoreBackend implementation. In addition, the CLIArgs trait must be implemented to map your command line arguments to the options and flags required by the general CLI implementation. After that, cli_main can be used to run the CLI interactively.

Macros

cli_write

Convenience macro for panicking if output fails.

cli_writeln

Convenience macro for panicking if output fails.

command

Create a Command from a help string and a function.

count

Structs

Command

A REPL command.

ListItem

Enums

KeyType

Traits

CLI

The interface required of a particular ledger-specific instantiation.

CLIArgs

CLI command line arguments.

CLIInput

Types which can be parsed from a string relative to a particular Keystore.Stream

Listable

Types which can be listed in terminal output and parsed from a list index.

Functions

cli_main

Run the CLI based in the provided command line arguments.

finish_transaction

key_gen

Type Definitions

CommandFunc

Keystore