Trait espresso_client::cli::CLI

pub trait CLI<'a> {
    type Ledger: 'static + Ledger;
    type Backend: 'a + KeystoreBackend<'a, Self::Ledger> + Send + Sync;
    type Loader: KeystoreLoader<Self::Ledger> + Send
    where
        <Self::Loader as KeystoreLoader<Self::Ledger>>::Meta == Self::Meta;
    type Meta: 'a + Send + Sync + Serialize + DeserializeOwned + Clone + PartialEq<Self::Meta>;
    type Args: CLIArgs;

    fn init_backend<'async_trait>(
        universal_param: &'a UniversalSrs<Bn<Parameters>>,
        args: Self::Args
    ) -> Pin<Box<dyn Future<Output = Result<Self::Backend, KeystoreError<Self::Ledger>>> + Send + 'async_trait, Global>>
    where
        'a: 'async_trait,
        Self: 'async_trait;
    fn init_loader<'async_trait>(
        storage: PathBuf,
        input: Reader
    ) -> Pin<Box<dyn Future<Output = Result<Self::Loader, KeystoreError<Self::Ledger>>> + Send + 'async_trait, Global>>
    where
        Self: 'async_trait;

    fn extra_commands() -> Vec<Command<'a, Self>, Global> { ... }
}

The interface required of a particular ledger-specific instantiation.

Required Associated Types

type Ledger: 'static + Ledger

The ledger for which we want to instantiate this CLI.

type Backend: 'a + KeystoreBackend<'a, Self::Ledger> + Send + Sync

The KeystoreBackend implementation to use for the keystore.

type Loader: KeystoreLoader<Self::Ledger> + Send
where
<Self::Loader as KeystoreLoader<Self::Ledger>>::Meta == Self::Meta

The KeystoreLoader implementation to use to create or load the keystore.

type Meta: 'a + Send + Sync + Serialize + DeserializeOwned + Clone + PartialEq<Self::Meta>

The type of metadata used by Self::Loader.

type Args: CLIArgs

The type of command line options for use when configuring the CLI.

Required Methods

fn init_backend<'async_trait>(
universal_param: &'a UniversalSrs<Bn<Parameters>>,
args: Self::Args
) -> Pin<Box<dyn Future<Output = Result<Self::Backend, KeystoreError<Self::Ledger>>> + Send + 'async_trait, Global>> where
'a: 'async_trait,
Self: 'async_trait,

Create a backend for the keystore which is being controlled by the CLI.

fn init_loader<'async_trait>(
storage: PathBuf,
input: Reader
) -> Pin<Box<dyn Future<Output = Result<Self::Loader, KeystoreError<Self::Ledger>>> + Send + 'async_trait, Global>> where
Self: 'async_trait,

Create a loader in order to create or load a new keystore for the CLI.

Provided Methods

fn extra_commands() -> Vec<Command<'a, Self>, Global>

Add extra, ledger-specific commands to the generic CLI interface.

This method is optional. By default it returns an empty list, in which case the CLI instantiation will still provide commands for all of the basic, ledger-agnostic keystore functionality.