Trait espresso_client::cli::CLIArgs

pub trait CLIArgs {
    fn key_gen_path(&self) -> Option<PathBuf>;
    fn storage_path(&self) -> Option<PathBuf>;
    fn io(&self) -> Option<SharedIO>;
    fn use_tmp_storage(&self) -> bool;
}

CLI command line arguments.

Required Methods

fn key_gen_path(&self) -> Option<PathBuf>

If specified, do not run the REPl, only generate a key pair in the given file.

fn storage_path(&self) -> Option<PathBuf>

Path to use for the keystore’s persistent storage.

If not provided, the default path, ~/.espresso/<ledger-name>/keystore, will be used.

fn io(&self) -> Option<SharedIO>

Override the default, terminal-based IO.

If this method returns Some, the CLI will use the provided IO adapters, without interactive line editing or password input hiding. Otherwise, it will use stdin (which must be a terminal) and stdout, with interactive line editing and password hiding.

fn use_tmp_storage(&self) -> bool

If true, create a temporary directory for storage instead of using CLIArgs::storage_path.