Struct espresso_client::jf_cap::structs::AssetPolicy

pub struct AssetPolicy {
    pub(crate) viewer_pk: ViewerPubKey,
    pub(crate) cred_pk: CredIssuerPubKey,
    pub(crate) freezer_pk: FreezerPubKey,
    pub(crate) reveal_map: RevealMap,
    pub(crate) reveal_threshold: Amount,
}

Policies related to asset record

viewer_pk - viewer public key

cred_pk - credential public key

freezer_pk - freezer public key

reveal_map - a binary vector indicating the subset of asset record info and identity attributes to be revealed to the viewer

Fields
viewer_pk: ViewerPubKeycred_pk: CredIssuerPubKeyfreezer_pk: FreezerPubKeyreveal_map: RevealMapreveal_threshold: Amount
Implementations

impl AssetPolicy

pub fn viewer_pub_key(&self) -> &ViewerPubKey

Reference to viewer public key

pub fn cred_creator_pub_key(&self) -> &CredIssuerPubKey

Reference to credential creator public key

pub fn freezer_pub_key(&self) -> &FreezerPubKey

Reference to freezer public key

pub fn reveal_map(&self) -> RevealMap

Getter for the reveal map USE WITH CAUTION: in most cases, you wouldn’t need to use reveal_map directly, try is_reveal_xxx_set() APIs first.

pub fn reveal_threshold(&self) -> Amount

Referecne to revealing threshold

pub fn set_reveal_map_for_test(self, reveal_map: RevealMap) -> AssetPolicy

Set the reveal map directly! USE WITH CAUTION!: since you might have an invalid reveal map, it is strongly recommended to construct reveal map via step-wise reveal_xxx() API.

pub fn set_reveal_threshold(self, reveal_threshold: Amount) -> AssetPolicy

Set revealing threshold policy

pub fn is_reveal_threshold_set(&self) -> bool

True if reveal_threshold is not the default value, false otherwise

pub fn set_viewer_pub_key(self, viewer_pub_key: ViewerPubKey) -> AssetPolicy

Set the viewer public key

pub fn is_viewer_pub_key_set(&self) -> bool

True if viewer public key is not the default key, false otherwise

pub fn set_cred_creator_pub_key(
self,
cred_creator_pub_key: CredIssuerPubKey
) -> AssetPolicy

Set the credential creator public key

pub fn is_cred_creator_pub_key_set(&self) -> bool

True if credential creator public key is not dummy, false otherwise

pub fn set_freezer_pub_key(self, freezer_pub_key: FreezerPubKey) -> AssetPolicy

Set the freezer public key

pub fn is_freezer_pub_key_set(&self) -> bool

True if freezer public key is not dummy, false otherwise

pub fn reveal_user_address(self) -> Result<AssetPolicy, TxnApiError>

Set policy to reveal user address to viewer Return TxnApiError::InvalidParameter if viewer public key has not been specified yet

pub fn is_user_address_revealed(&self) -> bool

Check if the policy is set to reveal user address

pub fn reveal_amount(self) -> Result<AssetPolicy, TxnApiError>

Set policy to reveal amount to viewer Return TxnApiError::InvalidParameter if viewer public key has not been specified yet

pub fn is_amount_revealed(&self) -> bool

Check if the policy is set to reveal amount

pub fn reveal_blinding_factor(self) -> Result<AssetPolicy, TxnApiError>

Set policy to reveal record commitment blinding factor to viewer Return TxnApiError::InvalidParameter if viewer public key has not been specified yet

pub fn is_blinding_factor_revealed(&self) -> bool

Check if the policy is set to reveal blinding factor

pub fn reveal_ith_attribute(self, i: usize) -> Result<AssetPolicy, TxnApiError>

Set policy to reveal ith identity attribute to viewer Return TxnApiError::InvalidParameter if viewer or credential creator public key have not been specified yet or it i greater or equal to ATTRS_LEN

pub fn reveal_all_attributes(self) -> Result<AssetPolicy, TxnApiError>

Set policy to reveal all identity attributes to viewer Return TxnApiError::InvalidParameter if viewer or credential creator public keys have not been specified yet

pub fn reveal_record_opening(self) -> Result<AssetPolicy, TxnApiError>

Set policy to reveal user address, amount and record commitment blinding factor to viewer Return TxnApiError::InvalidParameter if viewer public key have not been specified yet

pub fn reveal_all(self) -> Result<AssetPolicy, TxnApiError>

Set policy to reveal user address, amount, record commitment blinding factor and all identity attributes to viewer Return TxnApiError:: InvalidParameter if viewer or credential creator public keys have not been specified yet

impl AssetPolicy

pub fn rand_for_test<R>(rng: &mut R) -> AssetPolicy where
R: CryptoRng + RngCore,

generates a random AssetPolicy