Trait espresso_client::key_value_store::Index

pub trait Index<C> {
    fn insert(&mut self, change: C) -> Option<C>;
    fn remove(&mut self, change: &C) -> Option<C>;
}

An Interface for an in memory index which can insert and remove items To create a persistable index type you can implement this trait for whatever in memory index you choose to use. E.g. if you create a new type called CustomHashMap<K,V> you can implement Index<(K,V)> for it and then use the persitable type PersistableMap<(K,V), CustomHashMap<K,V>>.

Required Methods

fn insert(&mut self, change: C) -> Option<C>

Insert an item into the index. Returns the item value in the index before inserting if there was one Returns None if there was no item.

fn remove(&mut self, change: &C) -> Option<C>

Remove a value from the index. Returns the removed item if something was removed. None if the item was not in the index.