Module espresso_client::ledger_state

The ledger state module.

This module defines LedgerState and LedgerStates, which provide interfaces for building transactions and control the ledger state resource.

Structs

CryptoSnafu

SNAFU context selector for the TransactionError::CryptoError variant

FragmentationSnafu

SNAFU context selector for the TransactionError::Fragmentation variant

InsufficientBalanceSnafu

SNAFU context selector for the TransactionError::InsufficientBalance variant

InvalidFreezerKeySnafu

SNAFU context selector for the TransactionError::InvalidFreezerKey variant

InvalidSizeSnafu

SNAFU context selector for the TransactionError::InvalidSize variant

InvalidViewerKeySnafu

SNAFU context selector for the TransactionError::InvalidViewerKey variant

LedgerState

The state of the global ledger.

LedgerStates

Storage for the static and dynamic parts of the ledger state.

NoFitKeySnafu

SNAFU context selector for the TransactionError::NoFitKey variant

RecordAmount

Record amount.

TooManyOutputsSnafu

SNAFU context selector for the TransactionError::TooManyOutputs variant

TransactionUID

The UID of a transaction.

TransferSpec

Specifications for building a transfer.

Enums

ConvertRecordAmountError

The error that happens when converting a record amount.

TransactionError

Errors that happens during a transaction.

TransactionStatus

The status of a transaction.