Crate espresso_client::reef

An abstraction of a CAP-style network.

A network is CAP-style if it implements a UTXO-based blockchain where transactions are defined by nullifiers of the records they consume and commitments to the records they create. The network must support mint, transfer, and freeze transactions, as well as configurable asset privacy with viewing and freezing keys. It may optionally support additional operations that extend the CAP protocol. This library makes no assumptions about the detailed behavior of the network, such as the network architecture or the validation algorithm. It merely provides the types and traits needed by clients to interface with a CAP network.

Traits and types

This crate defines traits and types for working with a CAP-style network abstractly. It allows the development of network-agnostic components, by parameterizing such components on a type L which implements Ledger. These components can then be used with any compliant CAP-style network which implements these traits.

Minimal implementation

There is also a minimal implementation of the CAP ledger traits, in terms of Jellyfish CAP cryptographic primitives. The cap implementation can be used as a foundation for building more complex CAP ledgers. It can also be used as an easy mock implementation to test a network-agnostic component in isolation from any particular network.

Modules

cap

Minimal implementation of the ledger traits for a CAP-style ledger.

traits

Traits describing the interface of a CAP ledger.

types

Ledger types

Structs

ViewingMemoOpening

Information contained in a viewing memo.

Enums

ViewingError

Errors that can occur when trying to decrypt a viewing memo.

Traits

Ledger

A CAP ledger.

Type Definitions

Block

A block of transactions that can be applied to a ledger L.

BlockProof

A proof used to authenticate a committed block in a ledger L.

NullifierProof

A proof that a nullifier is spent or unspent, relative to a ledger L.

NullifierSet

A set of spent nullifiers for a ledger L.

StateCommitment

A commitment to a validator state for a ledger L.

Transaction

A transaction that can be applied to a ledger L.

TransactionHash

A committing hash of a transaction that can be applied to a ledger L.

TransactionKind

Types of transactions supported by a ledger L.

ValidationError

An error that can occur while validating transitions of a ledger L.

Validator

A validator for a ledger L.