Function espresso_client::reef::cap::open_xfr_viewing_memo

pub fn open_xfr_viewing_memo(
    assets: &HashMap<AssetCode, AssetDefinition, RandomState>,
    keys: &HashMap<ViewerPubKey, ViewerKeyPair, RandomState>,
    xfr: &TransferNote
) -> Result<ViewingMemoOpening, ViewingError>

Attempt to open the viewer memo attached to a CAP transfer transaction.

assets should be the set of asset types for which the caller holds the viewing key, indexed by asset code. This determines which asset types can be viewed by this method. keys is the caller’s collection of viewing key pairs, indexed by public key. keys must contain every public key which is listed as a viewer in the policy of one of the assets.