Trait espresso_client::reef::traits::TransactionKind

pub trait TransactionKind: Clone + Debug + Display + PartialEq<Self> + Eq + Hash + Serialize + DeserializeOwned + Send + Sync {
    fn send() -> Self;
    fn receive() -> Self;
    fn mint() -> Self;
    fn freeze() -> Self;
    fn unfreeze() -> Self;
    fn unknown() -> Self;
}

The types of transactions supported by this network.

This trait represents an enum-like interface, with a variant for each transaction. Thus, the public interface consists of constructors for the CAP transaction type variants. The implementation may include additional variants for extensions to the protocol. It is required that two transaction kinds representing different kinds of transactiosn compare unequal. For example, TransactionKind::send() != TransactionKind::mint().

Required Methods

fn send() -> Self

fn receive() -> Self

fn mint() -> Self

fn freeze() -> Self

fn unfreeze() -> Self

fn unknown() -> Self