Struct espresso_client::Keystore

pub struct Keystore<'a, Backend, L, Meta> where
    Backend: KeystoreBackend<'a, L>,
    L: 'static + Ledger,
    Meta: Serialize + DeserializeOwned + Send + Sync + Clone + PartialEq<Meta>, {
    pub(crate) mutex: Arc<KeystoreSharedStateRwLock<'a, L, Backend, Meta>>,
    pub(crate) task_scope: Scope<'a, (), AsyncStd>,
}

The generic CAP keystore implementation.

It is a soundness requirement that the destructor of a Keystore run when the Keystore is dropped. Therefore, std::mem::forget must not be used to forget a Keystore without running its destructor.

Fields
mutex: Arc<KeystoreSharedStateRwLock<'a, L, Backend, Meta>>task_scope: Scope<'a, (), AsyncStd>
Implementations

impl<'a, L, Backend, Meta> Keystore<'a, Backend, L, Meta> where
L: 'static + Ledger,
Backend: 'a + KeystoreBackend<'a, L> + Send + Sync,
Meta: 'a + Serialize + DeserializeOwned + Send + Sync + Clone + PartialEq<Meta>,

pub async fn create_stores(
loader: &mut impl KeystoreLoader<L, Meta = Meta>
) -> Result<KeystoreStores<'a, L, Meta>, KeystoreError<L>>

pub fn new<'l>(
backend: Backend,
loader: &'l mut impl KeystoreLoader<L, Meta = Meta> + Send
) -> Pin<Box<dyn Future<Output = Result<Keystore<'a, Backend, L, Meta>, KeystoreError<L>>> + Send + 'l, Global>> where
'a: 'l,

pub async fn write(
&self
) -> KeystoreSharedStateWriteGuard<'_, 'a, L, Backend, Meta>

Access the shared state directly.

pub async fn read(
&self
) -> RwLockReadGuard<'_, KeystoreSharedState<'a, L, Backend, Meta>>

Access the shared state directly.

pub async fn viewing_pub_keys(&self) -> Vec<ViewerPubKey, Global>

Get the viewing public keys.

pub async fn freezing_pub_keys(&self) -> Vec<FreezerPubKey, Global>

Get the freezing public keys.

pub async fn sending_addresses(&self) -> Vec<UserAddress, Global>

Get the sending addresses.

pub async fn sending_keys(&self) -> Vec<UserKeyPair, Global>

Get the sending keys.

pub async fn viewing_account(
&self,
pub_key: &ViewerPubKey
) -> Result<Account<L, ViewerKeyPair>, KeystoreError<L>>

Get the viewing account by the public key.

pub async fn freezing_account(
&self,
pub_key: &FreezerPubKey
) -> Result<Account<L, FreezerKeyPair>, KeystoreError<L>>

Get the freezing account by the public key.

pub async fn sending_account(
&self,
address: &UserAddress
) -> Result<Account<L, UserKeyPair>, KeystoreError<L>>

Get the sending account by the address.

pub async fn balance(&self, asset: &AssetCode) -> U256

Compute the spendable balance of the given asset owned by the given addresses.

pub async fn balance_breakdown(
&self,
address: &UserAddress,
asset: &AssetCode
) -> U256

Compute the spendable balance of the given asset owned by the given address.

pub async fn frozen_balance_breakdown(
&self,
address: &UserAddress,
asset: &AssetCode
) -> U256

Compute the balance frozen records of the given asset type owned by the given address.

pub async fn records(&self) -> Vec<Record, Global>

List records owned or viewable by this keystore.

pub async fn assets(&self) -> Vec<Asset, Global>

List assets discovered or imported by this keystore.

pub async fn asset(&self, code: AssetCode) -> Option<Asset>

Get details about an asset type using its code.

pub fn transaction_history(
&'l self
) -> Pin<Box<dyn SendFuture<'a, Result<Vec<Transaction<L>, Global>, KeystoreError<L>>, Output = Result<Vec<Transaction<L>, Global>, KeystoreError<L>>> + 'l, Global>>

List past transactions involving this keystore.

pub async fn transfer(
&mut self,
sender: Option<&UserAddress>,
asset: &AssetCode,
receivers: &[(UserPubKey, impl Clone + Into<RecordAmount>)],
fee: impl Into<RecordAmount>
) -> Result<TransactionUID<L>, KeystoreError<L>>

Basic transfer without customization.

To add transfer size requirement, call Keystore::build_transfer with a specified xfr_size_requirement.

sender

If provided, only this address will be used to transfer the asset.

Otherwise, all the owned addresses can be used for the transfer.

pub async fn build_transfer(
&mut self,
sender: Option<&UserAddress>,
asset: &AssetCode,
receivers: &[(UserPubKey, impl Clone + Into<RecordAmount>, bool)],
fee: impl Into<RecordAmount>,
bound_data: Vec<u8, Global>,
xfr_size_requirement: Option<(usize, usize)>
) -> Result<(TransferNote, TransactionParams<L>), KeystoreError<L>>

Build a transfer with full customization.

receivers: list of (receiver address, amount, burn)

pub async fn submit(
&mut self,
txn: <<<L as Ledger>::Validator as Validator>::Block as Block>::Transaction,
info: TransactionParams<L>
) -> Result<TransactionUID<L>, KeystoreError<L>>

Submit a transaction to be validated.

This function allows any kind of transaction to be submitted, even ledger-specific transaction types that are not part of the base CAP protocol.

pub async fn submit_cap(
&mut self,
txn: TransactionNote,
info: TransactionParams<L>
) -> Result<TransactionUID<L>, KeystoreError<L>>

Submit a CAP transaction to be validated.

pub async fn define_asset(
&mut self,
name: String,
description: &[u8],
policy: AssetPolicy
) -> Result<AssetDefinition, KeystoreError<L>>

Define a new asset and store secret info for minting.

pub async fn import_asset(
&mut self,
asset: Asset
) -> Result<(), KeystoreError<L>>

Import an asset.

Note that this function cannot be used to import verified assets. If the verified flag is set on asset, it will simply be ignored. Verified assets can only be imported using Keystore::verify_assets, conditional on a signature check.

pub async fn verify_assets(
&mut self,
trusted_signer: &VerKey<EdwardsParameters>,
library: VerifiedAssetLibrary
) -> Result<Vec<AssetDefinition, Global>, KeystoreError<L>>

Load a verified asset library from a file or byte stream.

trusted_signer must be the public key of an entity trusted by this application to verify assets. It must also be the public key which was used to sign library.

If successful, the asset definitions loaded from library are returned and their codes are added to this keystore’s set of verified asset codes. Note that assets loaded from a verified library are not persisted (unless the same assets are imported as unverified using Keystore::import_asset) in order to preserve the verified library as the single source of truth about verified assets. Therefore, this function must be called each time a keystore is created or opened in order to ensure that the verified assets show up in the keystore’s verified set.

pub fn add_viewing_account<'l>(
&'l mut self,
viewing_key: ViewerKeyPair,
description: String
) -> Pin<Box<dyn SendFuture<'a, Result<(), KeystoreError<L>>, Output = Result<(), KeystoreError<L>>> + 'l, Global>> where
'a: 'l,

Add a viewing key to the keystore’s key set.

pub fn generate_viewing_account<'l>(
&'l mut self,
description: String
) -> Pin<Box<dyn SendFuture<'a, Result<ViewerPubKey, KeystoreError<L>>, Output = Result<ViewerPubKey, KeystoreError<L>>> + 'l, Global>> where
'a: 'l,

Generate a new viewing key and add it to the keystore’s key set.

pub fn add_freezing_account<'l>(
&'l mut self,
freeze_key: FreezerKeyPair,
description: String
) -> Pin<Box<dyn SendFuture<'a, Result<(), KeystoreError<L>>, Output = Result<(), KeystoreError<L>>> + 'l, Global>> where
'a: 'l,

Add a freezing key to the keystore’s key set.

pub fn generate_freezing_account<'l>(
&'l mut self,
description: String
) -> Pin<Box<dyn SendFuture<'a, Result<FreezerPubKey, KeystoreError<L>>, Output = Result<FreezerPubKey, KeystoreError<L>>> + 'l, Global>> where
'a: 'l,

Generate a new freezing key and add it to the keystore’s key set.

pub fn add_sending_account<'l>(
&'l mut self,
user_key: UserKeyPair,
description: String,
scan_from: EventIndex
) -> Pin<Box<dyn SendFuture<'a, Result<(), KeystoreError<L>>, Output = Result<(), KeystoreError<L>>> + 'l, Global>> where
'a: 'l,

Add a sending key to the keystore’s key set.

Since this key was not generated by this keystore, it may have already been used and thus may own existing records. The keystore will start a scan of the ledger in the background to find records owned by this key. The scan will start from the event specified by scan_from.

pub fn generate_sending_account<'l>(
&'l mut self,
description: String,
scan_from: Option<EventIndex>
) -> Pin<Box<dyn SendFuture<'a, Result<UserPubKey, KeystoreError<L>>, Output = Result<UserPubKey, KeystoreError<L>>> + 'l, Global>> where
'a: 'l,

Generate a new sending key and add it to the keystore’s key set.

Keys are generated deterministically based on the mnemonic phrase used to load the keystore. If this is a recovery of an HD keystore from a mnemonic phrase, scan_from can be used to initiate a background scan of the ledger from the given event index to find records already belonging to the new key.

pub async fn import_memo(
&mut self,
memo: ReceiverMemo,
comm: RecordCommitment,
uid: u64,
proof: MerklePath<Fp256<FrParameters>>
) -> Result<(), KeystoreError<L>>

Manually add an encrypted record.

This can be used to access assets more quickly than waiting for a ledger scan when recovering a keystore.

pub async fn build_mint(
&mut self,
minter: Option<&UserAddress>,
fee: impl Into<RecordAmount>,
asset_code: &AssetCode,
amount: impl Into<RecordAmount>,
receiver: UserPubKey
) -> Result<(MintNote, TransactionParams<L>), KeystoreError<L>>

Create a mint note that assigns an asset to an owner.

pub async fn mint(
&mut self,
minter: Option<&UserAddress>,
fee: impl Into<RecordAmount>,
asset_code: &AssetCode,
amount: impl Into<RecordAmount>,
receiver: UserPubKey
) -> Result<TransactionUID<L>, KeystoreError<L>>

Build and submit a mint transaction.

See Keystore::build_mint.

pub async fn build_freeze(
&mut self,
freezer: Option<&UserAddress>,
fee: impl Into<RecordAmount>,
asset: &AssetCode,
amount: impl Into<U256>,
owner: UserAddress
) -> Result<(FreezeNote, TransactionParams<L>), KeystoreError<L>>

Build a transaction to freeze at least amount of a particular asset owned by a given user.

In order to freeze an asset, this keystore must be a viewer of that asset type, and it must have observed enough transactions to determine that the target user owns at least amount of that asset.

Freeze transactions do not currently support change, so the amount frozen will be at least amount but might be more, depending on the distribution of the freezable records we have for the target user.

Some of these restrictions will be rolled back in the future:

An API can be provided for freezing without being a viewer, if a freezable record opening is provided to us out of band by a viewer.

Keystore::build_freeze uses the same allocation scheme for input records as Keystore::build_transfer, which tries to minimize fragmentation. But freeze transactions do not increase fragmentation because they have no change output, so we could use a different allocation scheme that tries to minimize change, which would limit the amount we can over-freeze, and would guarantee that we freeze the exact amount if it is possible to make exact change with the freezable records we have.

pub async fn freeze(
&mut self,
freezer: Option<&UserAddress>,
fee: impl Into<RecordAmount>,
asset: &AssetCode,
amount: impl Into<U256>,
owner: UserAddress
) -> Result<TransactionUID<L>, KeystoreError<L>>

Build an submit a freeze transaction.

See Keystore::build_freeze.

pub async fn build_unfreeze(
&mut self,
freezer: Option<&UserAddress>,
fee: impl Into<RecordAmount>,
asset: &AssetCode,
amount: impl Into<U256>,
owner: UserAddress
) -> Result<(FreezeNote, TransactionParams<L>), KeystoreError<L>>

Build a transaction to unfreeze at least amount of a particular asset owned by a given user.

In order to unfreeze, this keystore must have previously been used to freeze at least amount of the target’s assets.

pub async fn unfreeze(
&mut self,
freezer: Option<&UserAddress>,
fee: impl Into<RecordAmount>,
asset: &AssetCode,
amount: impl Into<U256>,
owner: UserAddress
) -> Result<TransactionUID<L>, KeystoreError<L>>

Build and submit an unfreeze transaction.

See Keystore::build_unfreeze.

pub async fn transaction_status(
&self,
uid: &TransactionUID<L>
) -> Result<TransactionStatus, KeystoreError<L>>

Get the status of a transaction.

pub async fn await_transaction(
&self,
uid: &TransactionUID<L>
) -> Result<TransactionStatus, KeystoreError<L>>

A future which completes when the transaction is finalized (committed or rejected). Works only for transactions we submitted

pub async fn sync(&self, t: EventIndex) -> Result<(), KeystoreError<L>>

A future which completes when the keystore has processed events at least including t.

pub async fn now(&self) -> EventIndex

The index of the most recently processed event.

pub async fn sync_with_peer(
&self,
peer: &Keystore<'a, Backend, L, Meta>
) -> Result<(), KeystoreError<L>>

A future which completes when the keystore has processed at least as many events as peer.

pub async fn await_key_scan(
&self,
address: &UserAddress
) -> Result<(), KeystoreError<L>>

A future which completes when there are no more in-progress ledger scans for address.