Trait espresso_client::KeystoreBackend

pub trait KeystoreBackend<'a, L>: Send where
    L: Ledger, {
    type EventStream: 'a + Stream + Unpin + Send
    where
        <Self::EventStream as Stream>::Item == (LedgerEvent<L>, EventSource);

    fn create<'life0, 'async_trait>(
        &'life0 mut self
    ) -> Pin<Box<dyn Future<Output = Result<LedgerState<'a, L>, KeystoreError<L>>> + Send + 'async_trait, Global>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn subscribe<'life0, 'async_trait>(
        &'life0 self,
        from: EventIndex,
        to: Option<EventIndex>
    ) -> Pin<Box<dyn Future<Output = Self::EventStream> + Send + 'async_trait, Global>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn get_public_key<'life0, 'life1, 'async_trait>(
        &'life0 self,
        address: &'life1 UserAddress
    ) -> Pin<Box<dyn Future<Output = Result<UserPubKey, KeystoreError<L>>> + Send + 'async_trait, Global>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;
    fn register_user_key<'life0, 'life1, 'async_trait>(
        &'life0 mut self,
        pub_key: &'life1 UserKeyPair
    ) -> Pin<Box<dyn Future<Output = Result<(), KeystoreError<L>>> + Send + 'async_trait, Global>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;
    fn get_initial_scan_state<'life0, 'async_trait>(
        &'life0 self,
        from: EventIndex
    ) -> Pin<Box<dyn Future<Output = Result<(MerkleTree<Fp256<FrParameters>>, EventIndex), KeystoreError<L>>> + Send + 'async_trait, Global>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn get_nullifier_proof<'life0, 'life1, 'async_trait>(
        &'life0 self,
        block_height: u64,
        nullifiers: &'life1 mut <<<<L as Ledger>::Validator as Validator>::Block as Block>::Transaction as Transaction>::NullifierSet,
        nullifier: Nullifier
    ) -> Pin<Box<dyn Future<Output = Result<(bool, <<<<<L as Ledger>::Validator as Validator>::Block as Block>::Transaction as Transaction>::NullifierSet as NullifierSet>::Proof), KeystoreError<L>>> + Send + 'async_trait, Global>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;
    fn submit<'life0, 'async_trait>(
        &'life0 mut self,
        note: <<<L as Ledger>::Validator as Validator>::Block as Block>::Transaction,
        info: Transaction<L>
    ) -> Pin<Box<dyn Future<Output = Result<(), KeystoreError<L>>> + Send + 'async_trait, Global>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    fn finalize<'life0, 'async_trait>(
        &'life0 mut self,
        _txn: Transaction<L>,
        _txn_id: Option<(u64, u64)>
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait, Global>>
    where
        'life0: 'async_trait,
        L: 'static,
        Self: 'async_trait,
    { ... }
}

The interface required by the keystore from a specific network/ledger implementation.

This trait is the adapter for ledger-specific plugins into the ledger-agnostic Keystore implementation. It provides an interface for the ledger-agnostic keystore to communicate with remote network participants for a particular ledger. Implementing this trait for your specific ledger enables the use of the full generic Keystore interface with your ledger.

Required Associated Types

type EventStream: 'a + Stream + Unpin + Send
where
<Self::EventStream as Stream>::Item == (LedgerEvent<L>, EventSource)

Notifications about ledger state changes.

The keystore must be able to subscribe to a stream of events from the backend. These events occur whenever the ledger state changes, including new blocks being committed and owner memos being published. The backend may also include events when blocks are rejected. This is not required, but it can help the keystore report failures more quickly (if a rejection is not reported, the rejected transaction will eventually be considered timed out).

Each event yielded by the stream must be tagged with an EventSource indicating where the event came from. Implementations of KeystoreBackend may provide events from multiple sources (for example, block events from a network query service and memo events from a centralized memo store) or they may aggregate all events into a single stream from a single source.

Required Methods

fn create<'life0, 'async_trait>(
&'life0 mut self
) -> Pin<Box<dyn Future<Output = Result<LedgerState<'a, L>, KeystoreError<L>>> + Send + 'async_trait, Global>> where
'life0: 'async_trait,
Self: 'async_trait,

Create a new keystore.

This method should query the current state of the network (or at least some past state) and create a LedgerState consistent with that state. It must also persist the initial state by interacting with the storage layer directly.

fn subscribe<'life0, 'async_trait>(
&'life0 self,
from: EventIndex,
to: Option<EventIndex>
) -> Pin<Box<dyn Future<Output = Self::EventStream> + Send + 'async_trait, Global>> where
'life0: 'async_trait,
Self: 'async_trait,

Subscribe to the asynchronous ledgerevent stream.

fn get_public_key<'life0, 'life1, 'async_trait>(
&'life0 self,
address: &'life1 UserAddress
) -> Pin<Box<dyn Future<Output = Result<UserPubKey, KeystoreError<L>>> + Send + 'async_trait, Global>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Find the public encryption key associated with the given address.

Users in CAP are identified by addresses, but sending to a user also requires access to their public encryption key, for encrypting owner memos. This information must be published in some way, and accessed by the backend in this method.

fn register_user_key<'life0, 'life1, 'async_trait>(
&'life0 mut self,
pub_key: &'life1 UserKeyPair
) -> Pin<Box<dyn Future<Output = Result<(), KeystoreError<L>>> + Send + 'async_trait, Global>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Publish an encryption key so that other keystores can access it using KeystoreBackend::get_public_key.

fn get_initial_scan_state<'life0, 'async_trait>(
&'life0 self,
from: EventIndex
) -> Pin<Box<dyn Future<Output = Result<(MerkleTree<Fp256<FrParameters>>, EventIndex), KeystoreError<L>>> + Send + 'async_trait, Global>> where
'life0: 'async_trait,
Self: 'async_trait,

Get a ledger state from which to start a scan.

The backend must return the Merkle frontier at a time earlier than the from index, but it is not required to record the state at every event index, it may return an earlier Merkle frontier, such as the initial frontier. For this reason, the event index corresponding to the returned frontier is also returned.

fn get_nullifier_proof<'life0, 'life1, 'async_trait>(
&'life0 self,
block_height: u64,
nullifiers: &'life1 mut <<<<L as Ledger>::Validator as Validator>::Block as Block>::Transaction as Transaction>::NullifierSet,
nullifier: Nullifier
) -> Pin<Box<dyn Future<Output = Result<(bool, <<<<<L as Ledger>::Validator as Validator>::Block as Block>::Transaction as Transaction>::NullifierSet as NullifierSet>::Proof), KeystoreError<L>>> + Send + 'async_trait, Global>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Determine whether a nullifier is spent and obtain a proof.

nullifiers is a local cache of nullifiers. It may not contain a status or proof for every relevant nullifier. This function should check if nullifier is represented in nullifiers and query the network only if it is not. Optionally, this function may add the proof to the cache after obtaining a proof from the network.

block_height indicates the block height, or state number, for which nullifiers is the current nullifier set.

fn submit<'life0, 'async_trait>(
&'life0 mut self,
note: <<<L as Ledger>::Validator as Validator>::Block as Block>::Transaction,
info: Transaction<L>
) -> Pin<Box<dyn Future<Output = Result<(), KeystoreError<L>>> + Send + 'async_trait, Global>> where
'life0: 'async_trait,
Self: 'async_trait,

Submit a transaction to a validator.

Provided Methods

fn finalize<'life0, 'async_trait>(
&'life0 mut self,
_txn: Transaction<L>,
_txn_id: Option<(u64, u64)>
) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait, Global>> where
'life0: 'async_trait,
L: 'static,
Self: 'async_trait,

Record a finalized transaction.

If successful, txn_id contains the block ID and index of the committed transaction.

This function is optional and does nothing by default. The backend can override it to perform cleanup or post-processing on completed transactions.